backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.135.2",
    "httpx>=0.28.1",
    "openai>=2.30.0",
    "pydantic-settings>=2.13.1",
    "python-multipart>=0.0.22",
    "supabase>=2.28.3",
    "uvicorn[standard]>=0.42.0",
]
